refacto validate rows ban_ids
